+++ docs/.vitepress/theme/index.ts
@@ -1,5 +1,5 @@
-import DefaultTheme from 'vitepress/theme';
 import type { Theme } from 'vitepress';
+import DefaultTheme from 'vitepress/theme';
 import './style.css';
 
 export default {